backend/requirements.txt
fastapi
uvicorn
python-multipart
python-dotenv
fal-client
supabase
pydantic